Interface Tweaks
minor changes to the interface - notably: made the gui size larger to accommodate more elements. moved the keyboard to the bottom of the new window size; moved the editor down, made it grey (red was a bit loud and grey is closer-ish to the microbrute), added a combo box for each oscillator and positioned them in a sensible way. don't think the cbs listen for changes yet but visually i think that is an okay place for them to look.